# docs/swagger.yaml
basePath: /api
host: localhost:8000
info:
  contact: {}
  description: 'Сервіс з API, який дозволить: \n- дізнатись поточний курс біткоіну
    (BTC) у гривні (UAH) \n- підписати емейл на отримання інформації по зміні курсу
    \n- запит, який відправить всім підписаним користувачам актуальний курс.'
  title: GSES2 BTC application
  version: "1.0"
paths:
  /rate:
    get:
      description: Запит має повертати поточний курс BTC до UAH використовуючи будь-який
        third party сервіс з публічним API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - rate
  /sendEmails:
    post:
      description: Запит має отримувати актуальний курс BTC до UAH за допомогою third-party
        сервісу та відправляти його на всі електронні адреси, які були підписані раніше
      produces:
      - application/json
      responses:
        "200":
          description: E-mailʼи відправлено
          schema:
            type: string
        "500":
          description: Помилка на сервері
          schema:
            type: string
      tags:
      - subscription
  /subscribe:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Запит має перевірити, чи немає данної електронної адреси в поточній
        базі даних (файловій) і, в разі її відсутності, записувати її. Пізніше, за
        допомогою іншого запиту ми будемо відправляти лист на ті електронні адреси,
        які будуть в цій базі.
      parameters:
      - description: Електронна адреса, яку потрібно підписати
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email додано
          schema:
            type: string
        "409":
          description: Email вже є в базі даних
          schema:
            type: string
        "500":
          description: Помилка на сервері
          schema:
            type: string
      tags:
      - subscription
swagger: "2.0"